+++ src/app/_admin/roles-admin/roles-admin.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit, SimpleChanges } from '@angular/core';
 import { AdminService } from '../../_services/admin.service';
 import {FormGroup,FormControl,Validators} from '@angular/forms';
+import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
+import {MakeSureDialogComponent} from '../../_library/make-sure-dialog/make-sure-dialog.component';
+
 //Import all shared logic required for forms handling
 import {CustomValidators  } from '../../_helpers/custom.validators';
 import {Role} from "../../_models/role";
@@ -21,7 +24,7 @@ export class RolesAdminComponent implements OnInit {
   validation_messages = CustomValidators.getMessages();
   private _subscriptions : Subscription[] = new Array<Subscription>();
 
-  constructor(private adminService:AdminService) { }
+  constructor(private adminService:AdminService, public dialog: MatDialog) { }
 
   //Update html
   ngOnChanges(changes: SimpleChanges) {
@@ -73,14 +76,48 @@ export class RolesAdminComponent implements OnInit {
 
   deleteRole(id:number) {
     this.loading = true;
-    this._subscriptions.push(this.adminService.deleteRole(id).subscribe(result=>{
-      this.loading = false;
-      let role = this.roles.find(i => i.id == id);
-      let index = this.roles.indexOf(role);
-      this.roles.splice(index,1);      
-    }));    
-    console.log("Delete role : " + id);
+    let dialogRef = this.dialog.open(MakeSureDialogComponent, {
+      disableClose :true,
+      data:  {title: "Suprimer un role",
+              text:"Attention quand vous suprimez un role tous les utilizateurs avec ce role n'auront plus le role supprimé"
+            } 
+    });
+    this._subscriptions.push(dialogRef.afterClosed().subscribe((result : boolean) => {
+      if (result) {    
+        this._subscriptions.push(this.adminService.deleteRole(id).subscribe(result=>{
+          this.loading = false;
+          let role = this.roles.find(i => i.id == id);
+          let index = this.roles.indexOf(role);
+          this.roles.splice(index,1);      
+        }));    
+        console.log("Delete role : " + id);
+      } else {
+        this.loading = false;
+      }
+    }));
   }
+/*
+  openUserRemoveDialog() {
+    let dialogRef = this.dialog.open(MakeSureDialogComponent, {
+      disableClose :true,
+      data:  {title: "Suprimer votre compte",
+              text:"Attention cette operation est irreversible, vous allez effacer tous vos messages, vos documents..."
+            } 
+    });
+    this._subscriptions.push(dialogRef.afterClosed().subscribe((result : boolean) => {
+      if (result) {
+        this._subscriptions.push(this.userService.delete().subscribe(result => {
+          this.userService.setCurrent(new User(null));
+          User.removeToken();
+          this.router.navigate([""]); //Go back home
+        }));
+      }
+    }));
+  }
+*/
+
+
+
 
   ngOnDestroy() {    
     //Unsubscribe to all
